pyproject.toml
[project]
name = "danfe-generator"
version = "0.2.0"
description = "Gerador de DANFE (Documento Auxiliar da Nota Fiscal Eletrônica) personalizado"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Gabriel Ramos" }
]
keywords = ["danfe", "nfe", "nota-fiscal", "pdf", "brasil", "fiscal", "xml-generator", "sefaz"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
]

dependencies = [
    "brazilfiscalreport>=0.5.33",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
web = [
    "streamlit>=1.28.0",
    "watchdog>=3.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "types-PyYAML>=6.0.0",
    "black>=23.0.0",
    "pre-commit>=3.5.0",
    "pillow>=10.0.0",  # For logo tests
    "pytest-xdist>=3.0.0",
]
all = [
    "danfe-generator[web,dev]",
]

[project.scripts]
danfe = "danfe_generator.cli:main"

[project.urls]
Homepage = "https://github.com/gabrielramos/danfe-generator"
Documentation = "https://github.com/gabrielramos/danfe-generator#readme"
Repository = "https://github.com/gabrielramos/danfe-generator"
Issues = "https://github.com/gabrielramos/danfe-generator/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/danfe_generator"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src/danfe_generator"]
branch = true
omit = [
    "*/__pycache__/*",
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
show_missing = true
fail_under = 80

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["danfe_generator"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "brazilfiscalreport.*",
    "streamlit.*",
]
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
